+++ lib/main.dart
@@ -1,16 +1,10 @@
 import 'package:flutter/material.dart';
 
+import './task_manager.dart';
+
 void main() => runApp(MyApp());
 
-class MyApp extends StatefulWidget {
-  @override
-  State<StatefulWidget> createState() {
-    return _MyAppState();
-  }
-}
-
-class _MyAppState extends State<MyApp> {
-  List<String> _tasks = ['Surgery'];
+class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -18,22 +12,7 @@ class _MyAppState extends State<MyApp> {
         appBar: AppBar(
           title: Text('Scheduler'),
         ),
-        body: Column(
-          children: [
-            Container(
-              margin: EdgeInsets.all(10.0),
-              child: RaisedButton(
-                onPressed: () {
-                  setState(() {
-                    _tasks.add('task');
-                  });
-                },
-                child: Text('Create Task'),
-              ),
-            ),
-            ,
-          ],
-        ),
+        body: TaskManager(),
       ),
     );
   }
